requirements.txt
python-twitter
discord
